# lib/api-zod/src/generated/types/price.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Price {
  id: string;
  unitAmount?: number | null;
  currency: string;
  interval?: string | null;
  intervalCount?: number | null;
}