# lib/api-zod/src/generated/types/subscriptionStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SubscriptionStatus {
  hasActiveSubscription: boolean;
  subscriptionId?: string | null;
  planName?: string | null;
  status?: string | null;
  currentPeriodEnd?: string | null;
  cancelAtPeriodEnd?: boolean | null;
}